requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==22.2.0
brotlipy==0.7.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
conda==23.1.0
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-package-handling @ file:///croot/conda-package-handling_1672865015732/work
conda_package_streaming @ file:///croot/conda-package-streaming_1670508151586/work
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1673298753778/work
cycler==0.11.0
distlib==0.3.6
filelock==3.11.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.1.0
google-auth==2.17.0
google-auth-oauthlib==0.4.6
grpcio==1.53.0
hydra-core==1.3.1
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.1.0
importlib-resources==5.12.0
kiwisolver==1.4.4
lightly==1.2.47
lightly-utils==0.0.2
lightning-utilities==0.7.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.13.1
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
platformdirs==3.2.0
pluggy @ file:///tmp/build/80754af9/pluggy_1648024445381/work
protobuf==3.20.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
pytorch-lightning==1.9.3
pytz==2022.7.1
PyYAML==6.0
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.19.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
toolz @ file:///croot/toolz_1667464077321/work
torch==1.13.1
torchmetrics==0.11.1
torchvision==0.14.1
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
typing_extensions==4.5.0
ultralytics==8.0.74
urllib3 @ file:///croot/urllib3_1673575502006/work
virtualenv==20.21.0
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
zstandard @ file:///opt/conda/conda-bld/zstandard_1663827383994/work
